Split Fetch code
